requirements.txt
django
 djangorestframework
 environs
 gunicorn